+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 1.3.0
 
 * Getting Started and better Log in
 * 409 conflict on registration if email exists